+++ changelog.md
@@ -4,6 +4,9 @@
 
 ### Patch
 
+- Updated the [`eslint`](https://npm.im/eslint) peer dependency to `6 - 7`.
+- Updated the [`eslint-plugin-jsdoc`](https://npm.im/eslint-plugin-jsdoc) peer dependency to `15.11 - 25`.
+- Updated dependencies.
 - v13.0.6 changelog entry typo fix.
 
 ## 13.0.6